pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "himawari-dynamic-wallpaper"
version = "0.1.0"
description = "Cross-platform Himawari satellite dynamic wallpaper project"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
  { name = "rrmmkk29" }
]
keywords = ["himawari", "wallpaper", "satellite", "cross-platform", "playwright"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Win32 (MS Windows)",
  "Environment :: X11 Applications",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Desktop Environment",
  "Topic :: Multimedia :: Graphics",
]
dependencies = [
  "Pillow>=10.0.0",
  "playwright>=1.45.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]
release = [
  "pyinstaller>=6.0.0",
]

[project.scripts]
himawari-wallpaper = "himawari_wallpaper.cli:main"
himawari-wallpaper-gui = "himawari_wallpaper.gui:main"
himawari-wallpaper-cleanup = "himawari_wallpaper.uninstall:main"

[project.urls]
Homepage = "https://github.com/rrmmkk29/himawari8_wallpaper"
Repository = "https://github.com/rrmmkk29/himawari8_wallpaper.git"
Issues = "https://github.com/rrmmkk29/himawari8_wallpaper/issues"
Changelog = "https://github.com/rrmmkk29/himawari8_wallpaper/blob/main/CHANGELOG.md"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["himawari_wallpaper_webzoom"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
